// src/core/model/kix/faq/FAQArticleProperty.ts
export enum FAQArticleProperty {

    ID = 'ID',

    TITLE = 'Title',

    CATEGORY_ID = 'CategoryID',

    VISIBILITY = 'Visibility',

    LANGUAGE = 'Language',

    CONTENT_TYPE = 'ContentType',

    NUMBER = 'Number',

    KEYWORDS = 'Keywords',

    FIELD_1 = 'Field1',

    FIELD_2 = 'Field2',

    FIELD_3 = 'Field3',

    FIELD_4 = 'Field4',

    FIELD_5 = 'Field5',

    FIELD_6 = 'Field6',

    APPROVED = 'Approved',

    VALID_ID = 'ValidID',

    CREATED_BY = 'CreatedBy',

    CREATED = 'Created',

    CHANGED_BY = 'ChangedBy',

    CHANGED = 'Changed',

    ATTACHMENTS = 'Attachments',

    HISTORY = 'History',

    VOTES = 'Votes',

    LINK = 'LINK',

}